local/http_service/requirements.txt
fastapi
houston-client>=1.3.0